refactor(ui): move SessionCard to separate dev page
- Create new /dev/session-card page for SessionCard component showcase
- Remove SessionCard section from /dev/components page
- Add link to new session-card page in dev hub

The SessionCard component now has its own dedicated dev page showing:
- Different status variants (running, idle, new)
- Truncated long summaries
- List view layout
- Usage patterns with delete button
